refactor: definition
* allocate definition into its own class

* change imports to justify this change

* Add header docstring for file

Issue: #55
Tag: 2.0.0

Authored-by: Joshua Rose <joshuarose099@gmail.com>
