+++ urban/__main__.py
@@ -26,7 +26,7 @@ from urban_api import send_phrase_request
 from urban_commands import add_cols_argument
 from urban_commands import add_result_argument, add_word_argument
 from urban_commands import return_argument_parser
-from urban_utils import Definition, stringify_definition_tag
+from urban_definition import Definition
 
 
 def main():
